refactor: simplify project structure for binaries
